requirements.txt
streamlit
cohere
pyttsx3
python-dotenv
